% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AddMiscVDJData}
\alias{AddMiscVDJData}
\title{Adds VDJ data from lists to the misc slot}
\usage{
AddMiscVDJData(object, TCR = NULL, BCR = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{TCR}{list with TCR data}

\item{BCR}{list with BCR data}
}
\description{
Adds VDJ data from lists to the misc slot
}
