% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AddMissingVDJFamilies}
\alias{AddMissingVDJFamilies}
\title{Add missing VDJ families to a list of families.}
\usage{
AddMissingVDJFamilies(families)
}
\arguments{
\item{families}{Families to complete}
}
\description{
Add missing VDJ families to a list of families.
}
