% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{AddTreeMetadata}
\alias{AddTreeMetadata}
\title{Add metadata to phylogenetic tree}
\usage{
AddTreeMetadata(tree, object, metadata = NULL, features = NULL)
}
\arguments{
\item{tree}{Tree of class phylo}

\item{object}{Seurat object}

\item{metadata}{Metadata column(s) to add to the tree}

\item{features}{Feature(s) to add to the tree}
}
\description{
Add metadata to phylogenetic tree
}
