% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{AddVDJDataForAssay}
\alias{AddVDJDataForAssay}
\title{Add VDJ metadata to misc slot in object}
\usage{
AddVDJDataForAssay(
  assay,
  object,
  vdj.primary,
  vdj.secondary,
  vj.primary,
  vj.secondary,
  force = F
)
}
\arguments{
\item{assay}{VDJ assay type}

\item{object}{Seurat object}

\item{vdj.primary}{Data frame with the metadata columns for the primary heavy/alpha chains}

\item{vdj.secondary}{Data frame with the metadata columns for the secondary heavy/alpha chains}

\item{vj.primary}{Data frame with metadata columns for the primary light/beta chains}

\item{vj.secondary}{Data frame with the metadata columns for the secondary light/beta chains}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}
}
\description{
Add VDJ metadata to misc slot in object
}
