% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AppendVDJData}
\alias{AppendVDJData}
\title{Append VDJ data to an object}
\usage{
AppendVDJData(object, vdj.data, assay = c("BCR", "TCR"))
}
\arguments{
\item{object}{Seurat object to append the VDJ data to}

\item{vdj.data}{List with 4 dataframes (vdj.primary, vdj.secondary, vj.primary, vj.secondary) to append}

\item{assay}{VDJ assay to add this new data to}
}
\description{
Append VDJ data to an object
}
