% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{availableChains}
\alias{availableChains}
\title{Get available chains from object}
\usage{
availableChains(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Get available chains from object
}
