% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{AvailableChainsList}
\alias{AvailableChainsList}
\title{Create options list of chains for dropdown}
\usage{
AvailableChainsList(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Create options list of chains for dropdown
}
