% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{AvailableRegions}
\alias{AvailableRegions}
\title{Get regions for given chain}
\usage{
AvailableRegions(chain)
}
\arguments{
\item{chain}{One of H, L, A or B}
}
\description{
Get regions for given chain
}
