% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplot_vh}
\alias{barplot_vh}
\title{Barplot for IGHV-family per group}
\usage{
barplot_vh(
  object,
  group.by = NULL,
  groups.to.plot = NULL,
  region = c("V", "C"),
  chain = c("H", "L"),
  by.family = T
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{groups.to.plot}{Which groups of the group by column should be plotted. Default = NULL (all)}

\item{region}{Region to plot. Available options: 'V'(ariable) or 'C'(onstant)}

\item{chain}{Chain to plot. Options: 'H'(eavy) or 'L'(ight)}

\item{by.family}{Group genes of 1 family together. Default = TRUE}
}
\description{
Barplot for IGHV-family per group
}
