% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{BarplotClonotypes}
\alias{BarplotClonotypes}
\title{Barplot with clonotype distribution}
\usage{
BarplotClonotypes(
  object,
  group.by = NULL,
  subset = NULL,
  clonotypes = NULL,
  position = c("stack", "dodge")
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}

\item{clonotypes}{Clonotypes to plot. Default = top 10}

\item{position}{Position of the bars in the plots. Options = stack or dodge}
}
\description{
Barplot with clonotype distribution
}
