% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{BuildLineageTree}
\alias{BuildLineageTree}
\title{Build a lineage tree for the given sequences}
\usage{
BuildLineageTree(sequences, outgroup = NULL, root = F)
}
\arguments{
\item{sequences}{Named vector of sequences}

\item{outgroup}{Outgroup to root the tree}

\item{root}{Should the tree be rooted. Default = FALSE}
}
\description{
Build a lineage tree for the given sequences
}
