% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypes.R
\name{CalculateDistances.identity}
\alias{CalculateDistances.identity}
\title{Create distance matrix where identical sequences get following distances:
identical: 0
non-identical: 1}
\usage{
CalculateDistances.identity(sequences)
}
\arguments{
\item{sequences}{Vector of sequences}
}
\description{
Create distance matrix where identical sequences get following distances:
identical: 0
non-identical: 1
}
