% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypes.R
\name{CalculateDistances}
\alias{CalculateDistances}
\title{Calculate distances between sequences with the given distance metric}
\usage{
CalculateDistances(sequences, metric)
}
\arguments{
\item{sequences}{Vector of sequences}

\item{metric}{Distance metric to use}
}
\description{
Calculate distances between sequences with the given distance metric
}
