% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{CalculateDiversity.gini}
\alias{CalculateDiversity.gini}
\title{Calculate diversity using the gini index}
\usage{
CalculateDiversity.gini(frequencies)
}
\arguments{
\item{frequencies}{Vector of frequencies}
}
\description{
Calculate diversity using the gini index
}
