% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{CalculateDiversity.shannon}
\alias{CalculateDiversity.shannon}
\title{Calculate diversity using the shannon index}
\usage{
CalculateDiversity.shannon(frequencies)
}
\arguments{
\item{frequencies}{Vector of frequencies}
}
\description{
Calculate diversity using the shannon index
}
