% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDR3freq}
\alias{CDR3freq}
\title{Dimplot for IGXV-family}
\usage{
CDR3freq(
  object,
  columnname = c("l.cdr3", "h.cdr3", "cdr3"),
  group.by = "seurat_clusters"
)
}
\arguments{
\item{object}{Seurat object}

\item{columnname}{Chain to plot, available options:"l.cdr3","h.cdr3","cdr3" (=both)}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}
}
\description{
Dimplot for IGXV-family
}
