% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cdr3length}
\alias{cdr3length}
\title{Lineplot with the CDR3 lenght per group}
\usage{
cdr3length(object, group.by = NULL, subset = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}
}
\description{
Lineplot with the CDR3 lenght per group
}
