% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDR3Plot}
\alias{CDR3Plot}
\title{Plot length of CDR3 AA/NT sequences}
\usage{
CDR3Plot(
  object,
  group.by = NULL,
  subset = NULL,
  plot.type = c("ridge", "line"),
  sequence.type = c("AA", "NT"),
  color.theme = ColorThemes(),
  colors = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by.}

\item{subset}{Subset data to these groups}

\item{plot.type}{Type of plot. Options = ridge, line}

\item{sequence.type}{Which CDR3 sequence should be used? Options: "AA" or "NT"}

\item{color.theme}{Color theme to use. Default = DALI}

\item{colors}{Colors to use. This overwrites the selected color theme}
}
\description{
Plot length of CDR3 AA/NT sequences
}
