% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDR3Plot.ridge}
\alias{CDR3Plot.ridge}
\title{Plot length of CDR3 aa sequences as a ridgeplot}
\usage{
CDR3Plot.ridge(object, group.by, vdj.cdr3.column, vj.cdr3.column, cols)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by.}

\item{vdj.cdr3.column}{Column name for heavy/alpha cdr3 data}

\item{vj.cdr3.column}{Column name for light/beta cdr3 data}

\item{cols}{Colors to use}
}
\description{
Plot length of CDR3 aa sequences as a ridgeplot
}
