% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CircosPlot}
\alias{CircosPlot}
\title{Circosplot for family to gene distribution}
\usage{
CircosPlot(object, group.by = NULL, subset = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}
}
\description{
Circosplot for family to gene distribution
}
