% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ClearVDJ}
\alias{ClearVDJ}
\title{Remove the VDJ data from a seurat object}
\usage{
ClearVDJ(object, assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{VDJ assay to remove. Default = NULL (=> removes everything)}
}
\description{
Remove the VDJ data from a seurat object
}
