% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CloneConnGraph}
\alias{CloneConnGraph}
\title{Display connections between clusters}
\usage{
CloneConnGraph(
  object,
  reduction,
  group.by = NULL,
  groups.highlight = NULL,
  clonotype.column = NULL,
  color.theme = ColorThemes(),
  colors = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensionality reduction}

\item{group.by}{Metadata column to group the family data by.}

\item{groups.highlight}{Groups for which to highlight the edges in the graph}

\item{clonotype.column}{Metadata column with clonotype information. Default = clonotype}

\item{color.theme}{Color theme to use. Default = DALI}

\item{colors}{Colors to use. This overwrites the selected color theme}
}
\description{
Display connections between clusters
}
