% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ClonotypeFrequency.bar}
\alias{ClonotypeFrequency.bar}
\title{Plot the frequency of clonotypes in (subset of) clusters in a barplot}
\usage{
ClonotypeFrequency.bar(
  object,
  chains,
  group.by,
  use.sequence,
  sequence.type,
  clonotype.column,
  show.missing,
  bulk,
  threshold
)
}
\arguments{
\item{object}{Seurat object}

\item{chains}{Vector of chain(s) to plot}

\item{group.by}{Metadata column to group the family data by}

\item{use.sequence}{Use AA/NT sequence instead of clonotype}

\item{sequence.type}{What sequences to use, available options: "AA" or "NT"}

\item{clonotype.column}{Metadata column with clonotype information}

\item{show.missing}{Show missing values in plot}

\item{bulk}{Group all cells together and handle as bulk}
}
\description{
Plot the frequency of clonotypes in (subset of) clusters in a barplot
}
