% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ClonotypeFrequency}
\alias{ClonotypeFrequency}
\title{Plot the frequency of clonotypes in (subset of) clusters}
\usage{
ClonotypeFrequency(
  object,
  chain = availableChains(object),
  group.by = NULL,
  subset = NULL,
  use.sequence = F,
  sequence.type = c("AA", "NT"),
  clonotype.column = NULL,
  bulk = F,
  show.missing = F,
  legend = F
)
}
\arguments{
\item{object}{Seurat object}

\item{chain}{Chain to plot, available options: "L", "H", NULL (= both)}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}

\item{use.sequence}{Use AA/NT sequence instead of clonotype. Default = FALSE}

\item{sequence.type}{What sequences to use, available options: "AA" or "NT". Only functional when use.sequence = T}

\item{clonotype.column}{Metadata column with clonotype information. Default = 'clonotype'}

\item{bulk}{Group all cells together and handle as bulk. Default = FALSE}

\item{show.missing}{Show missing values in plot. Default = FALSE}

\item{legend}{Show legend. Default = FALSE}
}
\description{
Plot the frequency of clonotypes in (subset of) clusters
}
