% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ColorScale}
\alias{ColorScale}
\title{Return colorscales for heatmaps from name}
\usage{
ColorScale(
  name = c("coolwarm", "viridis", "gray to blue", "turning red"),
  n = 100
)
}
\arguments{
\item{name}{Colorscheme name}

\item{n}{Colors in spectrum. Default = 100}
}
\description{
Return colorscales for heatmaps from name
}
