% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{CreateVDJData}
\alias{CreateVDJData}
\title{Create new VDJ data from by merging original and new content together}
\usage{
CreateVDJData(new, orig = NULL)
}
\arguments{
\item{new}{New VDJ data}

\item{orig}{Original VDJ data}
}
\description{
Create new VDJ data from by merging original and new content together
}
