% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{DefaultAssayVDJ}
\alias{DefaultAssayVDJ}
\alias{DefaultAssayVDJ<-}
\title{Get and set the default VDJ assay}
\usage{
DefaultAssayVDJ(object, ...)

DefaultAssayVDJ(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{New VDJ assay}
}
\value{
The name of the default VDJ assay

An object with the new default VDJ assay
}
\description{
Get and set the default VDJ assay
}
