% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypes.R
\name{DefineClonotypes}
\alias{DefineClonotypes}
\title{Define clonotypes}
\usage{
DefineClonotypes(
  object,
  sequence = c("aa", "nt"),
  metric = c("identity"),
  column = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{sequence}{Use the amino acis or nucleotide sequence. Default = aa}

\item{metric}{Distance metric to use. Default = identity}

\item{column}{Column to add to metadata. Default = clonotype.<metric>}
}
\description{
Define clonotypes
}
