% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimPlot_vh}
\alias{DimPlot_vh}
\title{Dimplot for IGXV-family}
\usage{
DimPlot_vh(
  object,
  fam = NULL,
  chain = NULL,
  gene = NULL,
  reduction = "tsne",
  grid = T
)
}
\arguments{
\item{object}{Seurat object}

\item{fam}{Heavy or light chain family. Options are heavy or light, default = NULL}

\item{chain}{Heavy or light chain. Options are heavy/h or light/l, default = NULL
Is used in combination with gene.}

\item{gene}{c/j/d gene. Options are c, j or d, default = NULL
Is used in combination with chain.}

\item{reduction}{Specify which reduction to use, default = "tsne"}

\item{grid}{If TRUE, show per gene type in grid. If FALSE, show all genes types together on plot. Default=T}
}
\description{
Dimplot for IGXV-family
}
