% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimplotChainRegion}
\alias{DimplotChainRegion}
\title{Dimplot for IGXV-family}
\usage{
DimplotChainRegion(
  object,
  region = c("V", "D", "J", "C"),
  chain = AvailableChains(object),
  by.family = T,
  grid = T,
  highlight = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{region}{Region to plot. Available options: 'V'(ariable), 'D', 'J'(unction), 'C'(onstant).}

\item{chain}{Chain to plot. Options: 'H'(eavy), 'L'(ight) for BCR; 'A'(lpha), 'B'(eta) for TCR}

\item{by.family}{Group genes of 1 family together. Only effective with the V-gene. Default = TRUE}

\item{grid}{If TRUE, show per gene type in grid. If FALSE, show all genes types together on plot. Default = TRUE}

\item{highlight}{Family or gene to highlight. Default = NULL}

\item{...}{Extra parameters passed to Seurat::Dimplot}
}
\description{
Dimplot for IGXV-family
}
