% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypes.R
\name{DistToAdjacency}
\alias{DistToAdjacency}
\title{Convert a sparse distance matrix to an adjacency matrix}
\usage{
DistToAdjacency(dist.mat)
}
\arguments{
\item{dist.mat}{dgRMatrix with distance info}
}
\description{
Convert a sparse distance matrix to an adjacency matrix
}
