% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{EqualiseTableRows}
\alias{EqualiseTableRows}
\title{Return table X, with the same rows as table Y}
\usage{
EqualiseTableRows(plot_data_x, plot_data_y)
}
\arguments{
\item{plot_data_x}{table to be transformed}

\item{plot_data_y}{reference table}
}
\description{
Return table X, with the same rows as table Y
}
