% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ExpansionPlot}
\alias{ExpansionPlot}
\title{Plot of clonotype expansion}
\usage{
ExpansionPlot(
  object,
  reduction,
  clonotype.column = "clonotype",
  color.low = "lightgrey",
  color.mid = "red",
  color.high = "darkred",
  threshold = 1,
  positive.size = 0.5,
  negative.size = 0.5,
  positive.alpha = 1,
  negative.alpha = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Which dimensionality reduction to use}

\item{clonotype.column}{Metadata column with clonotype information. Default = 'clonotype'}

\item{color.low}{Color for cells without clonotype. Default = lightgrey}

\item{color.mid}{Color for cells with the medium number of clonotypes. Default = red}

\item{color.high}{Color for cells with the highest number of clonotypes. Default = darkred}

\item{threshold}{Cells with clonotype count < threshold are colored with the min.color. Default = 1}

\item{positive.size}{Size of the dots with clonotype count > threshold. Default = 0.5}

\item{negative.size}{Size of the dots with clonotype count <= threshold. Default = 0.5}

\item{positive.alpha}{Alpha of the dots with clonotype count > threshold. Default = 1}

\item{negative.alpha}{Alpha of the dots with clonotype count < threshold. Default = 1}
}
\description{
Plot of clonotype expansion
}
