% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FeaturePlotChainRegion}
\alias{FeaturePlotChainRegion}
\title{Featureplot of clonotypes}
\usage{
FeaturePlotChainRegion(object, clonotypes, clonotype.column = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{clonotypes}{Clonotypes to plot}

\item{clonotype.column}{Metadata column with clonotype information. Default = 'clonotype'}

\item{...}{Extra parameters to pass to Seurat::Dimplot}
}
\description{
Featureplot of clonotypes
}
