% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FieldForColumn}
\alias{FieldForColumn}
\title{Get field name for given column name}
\usage{
FieldForColumn(column, fields, columns)
}
\arguments{
\item{column}{Column for which you want the field}

\item{fields}{List of available fields}

\item{columns}{List of available columns}
}
\description{
Get field name for given column name
}
