% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{FormatDimred}
\alias{FormatDimred}
\title{Nicer formatting for the most common dimensionality reductions}
\usage{
FormatDimred(reduction)
}
\arguments{
\item{reduction}{Dimensionality reduction}
}
\description{
Nicer formatting for the most common dimensionality reductions
}
