% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetCategoricalColorPalette}
\alias{GetCategoricalColorPalette}
\title{Get a colorpalette for given categorical data}
\usage{
GetCategoricalColorPalette(data)
}
\arguments{
\item{data}{Categorical data}
}
\description{
Get a colorpalette for given categorical data
}
