% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetDataColumn}
\alias{GetDataColumn}
\title{Get metadata column for given chain and region}
\usage{
GetDataColumn(chain, region, by.family = F)
}
\arguments{
\item{chain}{VDJ chain}

\item{region}{VDJ region}

\item{by.family}{Data grouped by family. Default = F}
}
\description{
Get metadata column for given chain and region
}
