% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getDefaultVDJAssay}
\alias{getDefaultVDJAssay}
\title{Determine assay type from cellranger output}
\usage{
getDefaultVDJAssay(data.dir)
}
\arguments{
\item{data.dir}{Cellranger output directory}
}
\description{
Determine assay type from cellranger output
}
