% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetDimRedColors}
\alias{GetDimRedColors}
\title{Returns colorvector based on chosen theme and amount of categories}
\usage{
GetDimRedColors(theme = "DALI", n)
}
\arguments{
\item{theme}{coloscheme to use}

\item{n}{number of categories}
}
\description{
Returns colorvector based on chosen theme and amount of categories
}
