% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{GetDistance}
\alias{GetDistance}
\title{Get a distance matrix with distances calculated from a reference}
\usage{
GetDistance(sequences, distance.method = "lv")
}
\arguments{
\item{sequences}{Named vector of sequences}

\item{distance.method}{Method to use to calculate distance between sequences. Defaul = "lv" (Levenstein)}
}
\description{
Get a distance matrix with distances calculated from a reference
}
