% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{GetInfoForMetadata}
\alias{GetInfoForMetadata}
\title{Get metadata for the given assay/chain combination}
\usage{
GetInfoForMetadata(object, assay, chain)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{VDJ assay}

\item{chain}{vdj or vj chain}
}
\description{
Get metadata for the given assay/chain combination
}
