% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{GetMutationRate}
\alias{GetMutationRate}
\title{Returns the mutation rate of each cell's VDJ chain to "mutation.rate" metadata column}
\usage{
GetMutationRate(
  object,
  clonotype = NULL,
  regions = "V",
  reference,
  chain = c("VDJ", "VJ"),
  clonotype.column = "clonotype"
)
}
\arguments{
\item{object}{Seurat object#'}

\item{clonotype}{Clonotype to get mutation rate from. Default uses every clonotype available.}

\item{regions}{VDJ region to use sequence for. DEfault = V.}

\item{reference}{Path to fasta reference sequence file}

\item{chain}{Which chain to use}

\item{clonotype.column}{name of the metadata column that contains the clonotypes}
}
\description{
Returns the mutation rate of each cell's VDJ chain to "mutation.rate" metadata column
}
