% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{GetSequences}
\alias{GetSequences}
\title{Get sequences for reference fasta & clonotype}
\usage{
GetSequences(
  object,
  metadata,
  regions = "V",
  reference,
  chain = c("VDJ", "VJ")
)
}
\arguments{
\item{object}{Seurat object}

\item{metadata}{Metadata of the cells of one clonotype}

\item{regions}{VDJ regions to use sequence for. Default = V}

\item{reference}{path to the reference fasta file}

\item{chain}{Which chain to use?}
}
\description{
Get sequences for reference fasta & clonotype
}
