% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetVDJ_Dataframe}
\alias{GetVDJ_Dataframe}
\title{Construct a dataframe with vdj data for a seurat object}
\usage{
GetVDJ_Dataframe(data.dir, sequence.columns, use.filtered)
}
\arguments{
\item{data.dir}{directory containing the vdj data}

\item{sequence.columns}{sequence columns from an AIRR file}

\item{use.filtered}{Load filtered contig annotation. Default = TRUE}
}
\description{
Construct a dataframe with vdj data for a seurat object
}
