% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{GetVFamilies}
\alias{GetVFamilies}
\title{Extract V-family from a vector of v-genes}
\usage{
GetVFamilies(v_genes, assay)
}
\arguments{
\item{v_genes}{Vector of genes}

\item{assay}{VDJ assay type. Options = TCR; BCR}
}
\description{
Extract V-family from a vector of v-genes
}
