% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ggplotColors}
\alias{ggplotColors}
\title{Generate ggplot default colors}
\usage{
ggplotColors(n, h = c(15, 375))
}
\arguments{
\item{n}{Number of colors needed}

\item{h}{Minimum and maximum hue}
}
\description{
Generate ggplot default colors
}
