% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{Interactive_DALI}
\alias{Interactive_DALI}
\title{Interactive shiny application to explore a Seurat object (both with and without VDJ data)}
\usage{
Interactive_DALI(object = NULL)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Interactive shiny application to explore a Seurat object (both with and without VDJ data)
}
