% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{IsValidSeuratObject}
\alias{IsValidSeuratObject}
\title{Check validity of an object}
\usage{
IsValidSeuratObject(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Check validity of an object
}
