% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{LineageTree}
\alias{LineageTree}
\title{Build a phylogenetic tree based on v,d,j and/or c gene}
\usage{
LineageTree(
  object,
  clonotype,
  regions = "V",
  reference,
  chain = c("VDJ", "VJ"),
  clonotype.column = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{clonotype}{Clonotype to plot}

\item{regions}{VDJ regions to use sequence for. Default = V}

\item{reference}{Path to reference fasta. This file can either be found in the output of cellranger multi or can be the input reference file on which cellranger multi was ran.}

\item{chain}{Which chain to use?}

\item{clonotype.column}{Metadata column with clonotype information. Default = clonotype}
}
\description{
Build a phylogenetic tree based on v,d,j and/or c gene
}
