% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{LineageTreeVGene}
\alias{LineageTreeVGene}
\title{Create a phylogenetic tree of the v-gene sequence of either the heavy or light chain}
\usage{
LineageTreeVGene(
  object,
  clonotype,
  airr,
  reference,
  chain = c("VDJ", "VJ"),
  clonotype.column = NULL,
  do.plot = T
)
}
\arguments{
\item{object}{Seurat object}

\item{clonotype}{Clonotype to plot}

\item{airr}{Path to airr_rearrangement.tsv. This file can be found in cellranger multi output (> v6.1.1)}

\item{reference}{Path to reference fasta. This file can either be found in the output of cellranger multi or can be the input reference file on which cellranger multi was ran.}

\item{chain}{Which chain to use?}

\item{clonotype.column}{Metadata column with clonotype information. Default = clonotype}

\item{do.plot}{Should the tree be plotted. Default = TRUE}
}
\description{
Create a phylogenetic tree of the v-gene sequence of either the heavy or light chain
}
