% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{Parse10XVDJReferenceHeader}
\alias{Parse10XVDJReferenceHeader}
\title{Extract information from FASTA header of 10x VDJ reference}
\usage{
Parse10XVDJReferenceHeader(header)
}
\arguments{
\item{header}{Fasta header string}
}
\description{
Extract information from FASTA header of 10x VDJ reference
}
