% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSHM}
\alias{PlotSHM}
\title{Plots a Dimensionality reduction plot with cells colored by mutation rate}
\usage{
PlotSHM(
  object,
  clonotype = NULL,
  regions = "V",
  reference,
  chain = c("VDJ", "VJ"),
  clonotype.column = "clonotype",
  order = T
)
}
\arguments{
\item{object}{Seurat object}

\item{clonotype}{clonotype to plot its mutation rate. Default plots every clonotype}

\item{regions}{VDJ region to use sequence from.}

\item{reference}{path to reference fasta sequence file}

\item{chain}{Which chain to use}

\item{clonotype.column}{name of the metadata column that contains the clonotypes}

\item{order}{Plot positive cells in the foreground. Default = True}
}
\description{
Plots a Dimensionality reduction plot with cells colored by mutation rate
}
