% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Read10X_AIRR}
\alias{Read10X_AIRR}
\title{Load 10x VDJ data in a seurat object from 10X AIRR rearrangement tsv}
\usage{
Read10X_AIRR(object, file, assay, force = F)
}
\arguments{
\item{object}{Seurat object}

\item{file}{10X AIRR rearrangement tsv}

\item{assay}{VDJ assay for loaded data}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}
}
\description{
Load 10x VDJ data in a seurat object from 10X AIRR rearrangement tsv
}
