% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Read10X_multi_vdj}
\alias{Read10X_multi_vdj}
\title{Load 10x VDJ data into a Seurat object containing multiple samples merged together}
\usage{
Read10X_multi_vdj(
  object,
  data.dir,
  id.column = NULL,
  assay = NULL,
  force = F,
  sort.by = c("umis", "reads"),
  use.filtered = T,
  quiet = F
)
}
\arguments{
\item{object}{Seurat object}

\item{data.dir}{Cellranger output directory for every sample. Can be 1 directory or a named vector/list of multiple directories. The names must correspond to the sample ids in the metdata column specified by column.id}

\item{id.column}{Metadata column that contains the information to distinguish samples. Used to map the keys of the named vector to the sample in the object}

\item{assay}{VDJ assay type for loaded data. This is automatically detected from input, but can be overwritten when something goes wrong.}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}

\item{sort.by}{Column to sort the data to determine if chain is primary or secondary. Options = umis, reads}

\item{use.filtered}{Load filtered contig annotation. Default = TRUE}

\item{quiet}{Ignore warnings. Default = FALSE}
}
\description{
Load 10x VDJ data into a Seurat object containing multiple samples merged together
}
