% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Read10X_MultiVDJ}
\alias{Read10X_MultiVDJ}
\title{Load 10x VDJ data from multiple directories into a Seurat object containing multiple samples}
\usage{
Read10X_MultiVDJ(
  object,
  data.dir,
  id_column = NULL,
  assay = NULL,
  force = F,
  sort.by = c("umis", "reads"),
  use.filtered = T,
  quiet = F
)
}
\arguments{
\item{object}{Seurat object}

\item{data.dir}{directory containing Cellranger output directories for every sample. If column is specified this will be a named list}

\item{id_column}{(Optional:) Metadata column that contains the information to distinguish samples.}

\item{assay}{VDJ assay type for loaded data. This is automatically detected from input, but can be overwritten when something goes wrong.}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}

\item{sort.by}{Column to sort the data to determine if chain is primary or secondary. Options = umis, reads}

\item{use.filtered}{Load filtered contig annotation. Default = TRUE}

\item{quiet}{Ignore warnings. Default = FALSE}
}
\description{
Load 10x VDJ data from multiple directories into a Seurat object containing multiple samples
}
