% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Read10X_vdj}
\alias{Read10X_vdj}
\title{Load 10x VDJ data in a seurat object}
\usage{
Read10X_vdj(object, data.dir)
}
\arguments{
\item{object}{Seurat object}

\item{data.dir}{Cellranger output directory}
}
\description{
Load 10x VDJ data in a seurat object
}
