% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{Read10XVDJReference}
\alias{Read10XVDJReference}
\title{Read 10x VDJ reference FASTA and extract sequences for each gene}
\usage{
Read10XVDJReference(fasta)
}
\arguments{
\item{fasta}{Path to fasta reference}
}
\description{
Read 10x VDJ reference FASTA and extract sequences for each gene
}
