% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Read_AIRR}
\alias{Read_AIRR}
\title{Load VDJ data from an AIRR rearrangement file}
\usage{
Read_AIRR(
  object,
  file,
  type,
  fields,
  columns,
  only.productive = T,
  productive.field = "productive",
  force = F
)
}
\arguments{
\item{object}{Seurat object}

\item{file}{AIRR rearrangement tsv}

\item{type}{VDJ assay type for loaded data}

\item{fields}{Fields to keep from the AIRR file}

\item{columns}{Column names to map the field names to}

\item{only.productive}{Keep only productive rearrangements. Default = TRUE}

\item{productive.field}{Field containing productive information. Default = productive}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}
}
\description{
Load VDJ data from an AIRR rearrangement file
}
