% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ReadData}
\alias{ReadData}
\title{Load data in the Seurat object}
\usage{
ReadData(
  object,
  type,
  data,
  fields,
  columns = NULL,
  force = F,
  sort.by = c("umis", "reads")
)
}
\arguments{
\item{object}{Seurat object}

\item{type}{VDJ assay type for loaded data}

\item{data}{data frame containg all data}

\item{fields}{Fields to select from the data}

\item{columns}{Rename fields to these names. If not specified, just use the field names}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}

\item{sort.by}{Column to sort the data to determine if chain is primary or secondary. Options = umis, reads}
}
\description{
Load data in the Seurat object
}
