% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ShinyTrajectoryPlot}
\alias{ShinyTrajectoryPlot}
\title{Plotfunction to plot a simple Trajectory Analysis plot for the DALI shiny app}
\usage{
ShinyTrajectoryPlot(object, reduction, gene = 1, color.theme = "DALI")
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensionality reduction as basis for the analysis and plot.}

\item{gene}{Gene to base trajectory analysis on. Default = the first in the expression matrix}

\item{color.theme}{Colorscheme for the plot.}
}
\description{
Plotfunction to plot a simple Trajectory Analysis plot for the DALI shiny app
}
