% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SpectratypePlot}
\alias{SpectratypePlot}
\title{Plot length of CDR3 aa sequences}
\usage{
SpectratypePlot(
  object,
  group.by = NULL,
  subset = NULL,
  plot.type = c("ridge", "line"),
  sequence = c("aa", "nt")
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}

\item{plot.type}{Type of plot. Options = ridge, line}

\item{sequence}{Which CDR3 sequence should be used? Options: aa or nt}
}
\description{
Plot length of CDR3 aa sequences
}
