% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SpectratypePlot.ridge}
\alias{SpectratypePlot.ridge}
\title{Plot length of CDR3 aa sequences as a ridgeplot}
\usage{
SpectratypePlot.ridge(object, group.by, heavy.cdr3.column, light.cdr3.column)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{heavy.cdr3.column}{Column name for heavy cdr3 data}

\item{light.cdr3.column}{Column name for light cdr3 data}
}
\description{
Plot length of CDR3 aa sequences as a ridgeplot
}
