% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{SplitObject_VDJ}
\alias{SplitObject_VDJ}
\title{Splits a Seurat object into a list of objects with their respective VDJ data in the misc slot}
\usage{
SplitObject_VDJ(object, column)
}
\arguments{
\item{object}{Seurat object}

\item{column}{metadata column to split the data on}
}
\description{
Splits a Seurat object into a list of objects with their respective VDJ data in the misc slot
}
