% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SubsetAssay}
\alias{SubsetAssay}
\title{Helperfunction to subset BCR or TCR data in a Seurat object}
\usage{
SubsetAssay(object, assay = c("TCR", "BCR"))
}
\arguments{
\item{object}{Seurat object to be subsetted}

\item{assay}{BCR or TCR assay}
}
\description{
Helperfunction to subset BCR or TCR data in a Seurat object
}
