% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{SubsetObject_VDJ}
\alias{SubsetObject_VDJ}
\title{Subsets part of a Seurat Object specified by sample_id from a metadata column}
\usage{
SubsetObject_VDJ(object, subset.by, group.id, assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{subset.by}{Attribute for subsetting.}

\item{group.id}{Name of the group to subset for}

\item{assay}{VDJ assay (TCR or BCR) you want to keep. Default keeps both in the subset}
}
\description{
Subsets part of a Seurat Object specified by sample_id from a metadata column
}
