% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SubsetVDJData}
\alias{SubsetVDJData}
\title{Create a subset of VDJ data using given cellbarcodes}
\usage{
SubsetVDJData(object, barcodes, assay = c("TCR", "BCR"))
}
\arguments{
\item{object}{Seurat object to be split}

\item{barcodes}{barcodes of data to be retained}

\item{assay}{BCR or TCR data}
}
\description{
Create a subset of VDJ data using given cellbarcodes
}
