% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TrajectoryPlot}
\alias{TrajectoryPlot}
\title{Function to create a Trajectory analyses from from BCR or VDJ cell data using the dynoplot package}
\usage{
TrajectoryPlot(
  object,
  method,
  reduction,
  assay = NULL,
  min.occurence = 5,
  color.by = NULL,
  color.scheme = c("coolwarm", "viridis", "gray to blue", "turning red"),
  size_cells = 3,
  alpha_cells = 0.5
)
}
\arguments{
\item{object}{Seurat Object}

\item{method}{Method to calculate trajectory analyses. See https://dynverse.org/reference/dynmethods/ for options}

\item{reduction}{Dimensionality reduction for the plot}

\item{assay}{BCR or TCR assay. If not specified it will take all the data.}

\item{min.occurence}{Minimum occurence of a clonotype for it to be indicated on the graph. Default values = 5.}

\item{color.by}{Meta.data column to color the datapoints by. If empty a gradient will be used.}

\item{color.scheme}{Colorscheme of the colorgradient used to color by pseudotime when no color.by argument is provided.}

\item{size_cells}{Size of the plotted dots. Default size = 3.}

\item{alpha_cells}{Transparancy of the plotted dots. Default alpha = 0.5.}
}
\description{
Function to create a Trajectory analyses from from BCR or VDJ cell data using the dynoplot package
}
