% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{TranslateIMGTGappedSequence}
\alias{TranslateIMGTGappedSequence}
\title{Translate sequence which may contain gaps, indicate by .}
\usage{
TranslateIMGTGappedSequence(sequence)
}
\arguments{
\item{sequence}{sequence to translate}
}
\description{
Translate sequence which may contain gaps, indicate by .
}
