% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{TranslateIMGTGappedSequences}
\alias{TranslateIMGTGappedSequences}
\title{Translate sequences which may contain gaps, indicated by .}
\usage{
TranslateIMGTGappedSequences(sequences)
}
\arguments{
\item{sequences}{vector of sequences to translate}
}
\description{
Translate sequences which may contain gaps, indicated by .
}
