% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Uniqify_VDJ}
\alias{Uniqify_VDJ}
\title{Give cell barcodes and clonotypes from VDJ data an identifier to allow merging of VDJ data from multiple samples}
\usage{
Uniqify_VDJ(
  object,
  sample.suffix,
  assay = c("BCR", "TCR"),
  columns.to.update = c("clonotype", "barcode")
)
}
\arguments{
\item{object}{Seurat object}

\item{sample.suffix}{Suffix to add to the barcodes/clonotypes to make them unique}

\item{assay}{VDJ assay to make unique.}

\item{columns.to.update}{List of columns that should be suffixed with sample.suffix. Default = c("clonotype", "barcode")}
}
\description{
Give cell barcodes and clonotypes from VDJ data an identifier to allow merging of VDJ data from multiple samples
}
