% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{VDJDoublets}
\alias{VDJDoublets}
\title{Detect cells which have both TCR and BCR annotations and are therefore likely doublets}
\usage{
VDJDoublets(
  object,
  reduction = NULL,
  do.plot = T,
  return.object = T,
  metadata.column = "vdj.doublets"
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensionality reduction to use}

\item{do.plot}{Show a plot indicating the doublets}

\item{return.object}{Return the Seurat object with a column added with doublet annotation}

\item{metadata.column}{Metadata column to add with doublet information. Default = vdj.doublets}
}
\description{
Detect cells which have both TCR and BCR annotations and are therefore likely doublets
}
