% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VDJGraph}
\alias{VDJGraph}
\title{Display connections between clusters}
\usage{
VDJGraph(
  object,
  reduction,
  group.by = NULL,
  groups.highlight = NULL,
  clonotype.column = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensionality reduction}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{groups.highlight}{Groups for which to highlight the edges in the graph}

\item{clonotype.column}{Metadata column with clonotype information. Default = clonotype}
}
\description{
Display connections between clusters
}
