% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VolcanoPlotDEG}
\alias{VolcanoPlotDEG}
\title{Use Seurat's DEG Table to plot a volcano plot}
\usage{
VolcanoPlotDEG(data, c.scheme = c("coolwarm", "viridis"))
}
\arguments{
\item{data}{Table with DEG data from Seurat::Findmarkers()}

\item{c.scheme}{colorscheme to use}
}
\description{
Use Seurat's DEG Table to plot a volcano plot
}
