% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{AddVDJDataForType}
\alias{AddVDJDataForType}
\title{Add VDJ metadata to misc slot in object}
\usage{
AddVDJDataForType(type, object, heavy, light, force = F)
}
\arguments{
\item{type}{VDJ assay type}

\item{object}{Seurat object}

\item{heavy}{Data frame with the metadata columns for the heavy chains}

\item{light}{Data frame with the metadata columns for the light chains}

\item{force}{Add VDJ data without checking overlap in cell-barcodes. Default = FALSE}
}
\description{
Add VDJ metadata to misc slot in object
}
