% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{AvailableChains}
\alias{AvailableChains}
\title{Get available chains from object}
\usage{
AvailableChains(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Get available chains from object
}
