% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{AvailableRegions}
\alias{AvailableRegions}
\title{Get regions for given chain}
\usage{
AvailableRegions(chain)
}
\arguments{
\item{chain}{VDJ (heavy/alpha) OR VJ (light/beta)}
}
\description{
Get regions for given chain
}
