% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplot_vh}
\alias{barplot_vh}
\title{Barplot for IGHV-family per group}
\usage{
barplot_vh(object, group.by = NULL, chain = "h")
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{chain}{Chain to plot. Can be any of c('l', 'light', 'h', 'heavy')}
}
\description{
Barplot for IGHV-family per group
}
