% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplot_vh}
\alias{barplot_vh}
\title{Barplot for IGHV-family per group}
\usage{
barplot_vh(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  region = c("V", "C"),
  chain = availableChains(object),
  by.family = T,
  legend = T,
  grid = F,
  add.missing.families = T,
  percent.total = T,
  show.missing.values = F
)
}
\arguments{
\item{object}{Seurat object}

\item{ident.1}{Identy class(es) to plot}

\item{ident.2}{Second identity class(es). This class will be used as comparison. If NULL, ident.1 will just be used to subset the data}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{region}{Region to plot. Available options: 'V'(ariable) or 'C'(onstant)}

\item{chain}{Chain to plot. Options: 'H'(eavy), 'L'(ight) for BCR; 'A'(lpha), 'B'(eta) for TCR}

\item{by.family}{Group genes of 1 family together. Default = TRUE}

\item{legend}{Should the legend be included in the plot. Default = TRUE}

\item{grid}{Organize plots in grid. Each plot contains information about 1 group. Default = FALSE}

\item{add.missing.families}{Should missing families be added to the plot. Default = TRUE}

\item{percent.total}{Should the fraction of cells be calculated from the total number or cells in the group or just the cells with VDJ info. Default = TRUE (= from total)}

\item{show.missing.values}{Should missing values be shown in the plot. Default = FALSE}
}
\description{
Barplot for IGHV-family per group
}
