% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{BarplotClonotypes}
\alias{BarplotClonotypes}
\title{Barplot with clonotype distribution}
\usage{
BarplotClonotypes(
  object,
  group.by = NULL,
  subset = NULL,
  clonotypes = NULL,
  position = c("stack", "dodge"),
  color.theme = ColorThemes(),
  colors = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by.}

\item{subset}{Subset data to these groups}

\item{clonotypes}{Clonotypes to plot. Default = top 10}

\item{position}{Position of the bars in the plots. Options = stack or dodge}

\item{color.theme}{Color theme to use. Default = DALI}

\item{colors}{Colors to use. This overwrites the selected color theme}
}
\description{
Barplot with clonotype distribution
}
