% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculateFrequency}
\alias{calculateFrequency}
\title{Calculate frequencies of column and group}
\usage{
calculateFrequency(object, data.column, group.by, show.missing)
}
\arguments{
\item{object}{Seurat object}

\item{data.column}{Metadata column to use}

\item{group.by}{Metadata column with group info}

\item{show.missing}{Should missing values be included or not}
}
\description{
Calculate frequencies of column and group
}
