% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDR3freq}
\alias{CDR3freq}
\title{Dimplot for IGXV-family}
\usage{
CDR3freq(object, chain = c("L", "H"), group.by = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{chain}{Chain to plot, available options: "L", "H", NULL (= both)}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}
}
\description{
Dimplot for IGXV-family
}
