% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cdr3length}
\alias{cdr3length}
\title{Plot length of CDR3 aa sequences}
\usage{
cdr3length(
  object,
  group.by = NULL,
  subset = NULL,
  plot.type = c("ridge", "line")
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{subset}{Subset data to these groups}

\item{plot.type}{Type of plot. Options = ridge, line}
}
\description{
Plot length of CDR3 aa sequences
}
