% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CDR3Plot.line}
\alias{CDR3Plot.line}
\title{Plot length of CDR3 aa sequences as a line plot}
\usage{
CDR3Plot.line(
  object,
  group.by,
  heavy.cdr3.column,
  light.cdr3.column,
  sequence.type
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by. Default = seurat_clusters}

\item{heavy.cdr3.column}{Column name for heavy cdr3 data}

\item{light.cdr3.column}{Column name for light cdr3 data}

\item{sequence.type}{AA or NT}
}
\description{
Plot length of CDR3 aa sequences as a line plot
}
