% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CircosPlotChains}
\alias{CircosPlotChains}
\title{Circosplot for VDJ chain useage distribution}
\usage{
CircosPlotChains(
  object,
  group.by = NULL,
  subset = NULL,
  color.theme = ColorThemes()
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Metadata column to group the family data by.}

\item{subset}{Subset data to these groups}

\item{color.theme}{Color theme to use. Default = DALI}
}
\description{
Circosplot for VDJ chain useage distribution
}
