% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ClonotypeDiversity}
\alias{ClonotypeDiversity}
\title{Calculate diversity in VDJ data}
\usage{
ClonotypeDiversity(
  object,
  algorithm = c("gini", "shannon"),
  use.sequence = F,
  sequence.type = c("AA", "NT"),
  chain = availableChains(object),
  clonotype.column = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{algorithm}{Algorithm used for diversity calculation. Options = "gini" or "shannon"}

\item{use.sequence}{Use AA/NT sequence instead of clonotype. Default = FALSE}

\item{sequence.type}{What sequences to use, available options: "AA" or "NT". Only functional when use.sequence = T}

\item{chain}{Chain to plot. Options: 'H'(eavy), 'L'(ight) for BCR; 'A'(lpha), 'B'(eta) for TCR}

\item{clonotype.column}{Metadata column with clonotype information. Default = 'clonotype'}
}
\description{
Calculate diversity in VDJ data
}
