% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ClonotypeToSequence}
\alias{ClonotypeToSequence}
\title{Convert clonotype to sequences}
\usage{
ClonotypeToSequence(
  object,
  ct,
  chain,
  clonotype.column = "clonotype",
  sequence.type = "AA",
  collapse = "\\n"
)
}
\arguments{
\item{object}{Seurat object}

\item{ct}{clonotype to convert}

\item{chain}{Which chain to get sequence for}

\item{clonotype.column}{Metadata column with clonotype information. Default = clonotype}

\item{sequence.type}{One of AA or NT. Default = AA}

\item{collapse}{Character to use to paste multiple sequence together}
}
\description{
Convert clonotype to sequences
}
