% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimPlot_vh}
\alias{DimPlot_vh}
\title{Dimplot for IGXV-family}
\usage{
DimPlot_vh(
  object,
  chain = c("H", "L", "A", "B"),
  region = c("V", "D", "J", "C"),
  by.family = T,
  grid = T,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{chain}{Chain to plot. Options: 'H'(eavy) or 'L'(ight)}

\item{region}{Region to plot. Available options: 'V'(ariable), 'D', 'J'(unction), 'C'(onstant),}

\item{by.family}{Group genes of 1 family together. Only effective with the V-gene. Default = TRUE}

\item{grid}{If TRUE, show per gene type in grid. If FALSE, show all genes types together on plot. Default = TRUE}

\item{...}{Extra parameters passed to Seurat::Dimplot}
}
\description{
Dimplot for IGXV-family
}
