% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FeaturePlot_vdj}
\alias{FeaturePlot_vdj}
\title{Featureplot of clonotypes}
\usage{
FeaturePlot_vdj(object, clonotypes, clonotype.column = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{clonotypes}{Clonotypes to plot}

\item{clonotype.column}{Column in meta data with clonotype info. Default = clonotype}

\item{...}{Extra parameters to pass to Seurat::Dimplot}
}
\description{
Featureplot of clonotypes
}
