% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{formatDimred}
\alias{formatDimred}
\title{Nicer formatting for the most common dimensionality reductions}
\usage{
formatDimred(reduction)
}
\arguments{
\item{reduction}{Dimensionality reduction}
}
\description{
Nicer formatting for the most common dimensionality reductions
}
