% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_v_families}
\alias{get_v_families}
\title{Extract V-family from a vector of v-genes}
\usage{
get_v_families(v_genes)
}
\arguments{
\item{v_genes}{Vector of genes}
}
\description{
Extract V-family from a vector of v-genes
}
