% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetAIRRSequenceColumns}
\alias{GetAIRRSequenceColumns}
\title{Get The sequence columns from an airr_rearrangement.tsv file}
\usage{
GetAIRRSequenceColumns(data.dir)
}
\arguments{
\item{data.dir}{path to directory containing the airr_rearrangement.tsv file}
}
\description{
Get The sequence columns from an airr_rearrangement.tsv file
}
