% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getAssayForData}
\alias{getAssayForData}
\title{Determine assay type from metrics_summary.csv}
\usage{
getAssayForData(csv.path)
}
\arguments{
\item{csv.path}{Path to metrics_summary.csv}
}
\description{
Determine assay type from metrics_summary.csv
}
