% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetCategoricalColorPalette}
\alias{GetCategoricalColorPalette}
\title{Returns vector of colors based on the selected color theme}
\usage{
GetCategoricalColorPalette(data, theme = ColorThemes())
}
\arguments{
\item{data}{Data to get categories from}

\item{theme}{Colorscheme to use. Default = DALI}
}
\description{
Returns vector of colors based on the selected color theme
}
