% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetCategoricalColorPalette}
\alias{GetCategoricalColorPalette}
\title{Returns colorvector based on chosen theme and amount of categories}
\usage{
GetCategoricalColorPalette(data, theme = "DALI")
}
\arguments{
\item{data}{data to get categories from}

\item{theme}{coloscheme to use. Default = DALI}
}
\description{
Returns colorvector based on chosen theme and amount of categories
}
