% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetSequence}
\alias{GetSequence}
\title{Get the sequence of a v,d,j or c gene for a cell}
\usage{
GetSequence(
  object,
  cell,
  assay = NULL,
  chain = c("VDJ", "VJ"),
  region = c("V", "D", "J", "C")
)
}
\arguments{
\item{object}{Seurat object}

\item{cell}{Cell id}

\item{assay}{VDJ assay to use. Default = DefaultAssayVDJ(object)}

\item{chain}{Chain to use. Options = VDJ (=heavy/alpha); VJ (=light/beta)}

\item{region}{Region to get sequence from}
}
\description{
Get the sequence of a v,d,j or c gene for a cell
}
