% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{Interactive_VDJ}
\alias{Interactive_VDJ}
\title{Explore VDJ data in an interactive way}
\usage{
Interactive_VDJ(object = NULL, vdj = T)
}
\arguments{
\item{object}{Seurat object}

\item{vdj}{Boolean to specify if object contains VDJ data. Default = True}
}
\description{
Explore VDJ data in an interactive way
}
