% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{Interactive_VDJ}
\alias{Interactive_VDJ}
\title{Explore VDJ data in an interactive way}
\usage{
Interactive_VDJ(object = NULL, max.upload.size = 1000)
}
\arguments{
\item{object}{Seurat object}

\item{max.upload.size}{Maximum upload size for shiny application (in MB). Default = 1000}
}
\description{
Explore VDJ data in an interactive way
}
