% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{isValidSeuratObject}
\alias{isValidSeuratObject}
\title{Check validity of an object}
\usage{
isValidSeuratObject(object)
}
\arguments{
\item{object}{Seurat object}
}
\description{
Check validity of an object
}
