% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{LineageTree}
\alias{LineageTree}
\title{Build a lineage tree for the given sequences}
\usage{
LineageTree(sequences, distance.method = "lv", outgroup = NULL, root = F)
}
\arguments{
\item{sequences}{Named vector of sequences}

\item{distance.method}{Method to use to compute distances between the sequences. Default = lv (Levenshtein)}

\item{outgroup}{Outgroup to root the tree}

\item{root}{Should the tree be rooted. Default = FALSE}
}
\description{
Build a lineage tree for the given sequences
}
