% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{MergeVDJ}
\alias{MergeVDJ}
\title{Merge seuratobjects with including their VDJ data}
\usage{
MergeVDJ(...)
}
\arguments{
\item{...}{seurat objects separated by comma}
}
\description{
Merge seuratobjects with including their VDJ data
}
