% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{MergeVDJ}
\alias{MergeVDJ}
\title{Merge multiple seurat objects and include their VDJ data}
\usage{
MergeVDJ(...)
}
\arguments{
\item{...}{Seurat objects separated by comma}
}
\description{
Merge multiple seurat objects and include their VDJ data
}
