% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_expansion}
\alias{plot_expansion}
\title{Plot of clonotype expansion}
\usage{
plot_expansion(
  object,
  reduction,
  clonotype.column = "clonotype",
  min.color = NULL,
  max.color = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Which dimensionality reduction to use}

\item{clonotype.column}{Metadata column with clonotype information. Default = 'clonotype'}

\item{min.color}{Color for cells without clonotype. Default = grey}

\item{max.color}{Color for cells with the highest number of clonotypes. Default = red}
}
\description{
Plot of clonotype expansion
}
