% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{SplitObject_VDJ}
\alias{SplitObject_VDJ}
\title{Splits a Seurat object into a list of objects with their respective VDJ data in the misc slot}
\usage{
SplitObject_VDJ(object, split.by, quiet = F)
}
\arguments{
\item{object}{Seurat object}

\item{split.by}{Attribute for splitting.}

\item{quiet}{Supress warnings. Default = FALSE}
}
\description{
Splits a Seurat object into a list of objects with their respective VDJ data in the misc slot
}
