% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{SubsetObject_VDJ}
\alias{SubsetObject_VDJ}
\title{Subsets part of a Seurat Object specified by sample_id from a metadata column}
\usage{
SubsetObject_VDJ(object, id_column, sample_id, assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{id_column}{metadata column containing the sample identifier}

\item{sample_id}{id in the id_column that corresponds to the to be subsetted sample}

\item{assay}{assay (TCR or BCR) you want to keep. Default keeps both in the subset}
}
\description{
Subsets part of a Seurat Object specified by sample_id from a metadata column
}
