% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Uniqify_VDJ}
\alias{Uniqify_VDJ}
\title{Give cellbarcodes and clonotypes from VDJ data an identifier}
\usage{
Uniqify_VDJ(object, data, assay = c("BCR", "TCR"), index)
}
\arguments{
\item{object}{Seurat object}

\item{data}{list of vdj dataframes}

\item{assay}{specify assay}

\item{index}{object name/id}
}
\description{
Give cellbarcodes and clonotypes from VDJ data an identifier
}
